+++ automation-scripts/6-org-policies/6-org-policies.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 set -xe
 
-ls -la
-rm -rf -- $(ls | grep -v env.tar.gz)
-ls -la
-tar -zxf env.tar.gz
-ls -la
-rm -f env.tar.gz
+if [ -f "./env.tar.gz" ]; then
+  ls -la
+  rm -rf -- $(ls | grep -v env.tar.gz)
+  ls -la
+  tar -zxf env.tar.gz
+  ls -la
+  rm -f env.tar.gz
+fi
 
 base_dir=$(pwd)
 
@@ -17,7 +19,7 @@ chmod 755 ./tf-wrapper.sh
 
 #get organization_id
 export ORGANIZATION_ID=$(terraform -chdir="../0-bootstrap/" output -json common_config | jq '.org_id' --raw-output)
-gcloud scc notifications describe "scc-notify" --organization=${ORGANIZATION_ID}
+$(gcloud scc notifications describe "scc-notify" --organization=${ORGANIZATION_ID} ) || true
 
 #Retrieve Service Account Email
 export GOOGLE_IMPERSONATE_SERVICE_ACCOUNT=$(terraform -chdir="../0-bootstrap/" output -raw organization_step_terraform_service_account_email)
@@ -93,9 +95,15 @@ pwd
 set +e
 unset GOOGLE_IMPERSONATE_SERVICE_ACCOUNT
 
+rm -rf ./development/.terraform
+rm -rf ./nonproduction/.terraform
+rm -rf ./production/.terraform
+rm -rf ./management/.terraform
+rm -rf ./identity/.terraform
+rm -rf ./common/.terraform
 cd ..
 pwd
 tar -zcf env.tar.gz --exclude env.tar.gz . 
+tar -zcf env.tar.gz --exclude env.tar.gz --exclude .git --exclude docs . 
 ls -la
 pwd
-
